⚡️ :: update rank entity
