+++ domain/src/main/java/com/semicolon/domain/entity/rank/SchoolRankAndSearchEntity.kt
@@ -1,13 +1,14 @@
 package com.semicolon.domain.entity.rank
 
-data class SearchSchoolEntity(
-    val schoolList: List<SchoolInfo>
+data class SchoolRankAndSearchEntity(
+    val schoolRankList: List<SchoolRank>
 ) {
-    data class SchoolInfo(
+    data class SchoolRank(
         val schoolId: Int,
         val schoolName: String,
         val ranking: Int,
         val logoImageUrl: String,
-        val walkCount: Int
+        val walkCount: Int,
+        val userCount: Int
     )
 }
